src/requirements.txt
GitPython==3.1.40
requests==2.26.0
pygount
pandas
scikit-learn
joblib
